Remove different timeframe for Masternode Polls
For simplicity, all proposal types should stay on the same 30 day timeframe. Generally the advantage of quicker decisions is outweighed by the disadvantage of a more complex governance system overall, with multiple timeframes, both in code and socially.

For example, most MNO only pay attention once a month, if governance polls were on a separate timeframe they would be settled before many MNO even saw them.

Also the advantage of quicker decisions applies equally to funding proposals as it does to decision pools, but 30 days is what we have and are used to. Unless it’s a problem, it should not be changed. 

There is nothing special about decision polls that justify the increased complexity of implementing a separate governance timeframe for them.